Se corrigen algunos errores
